API

Krotov.KrotovResult

Krotov.optimize_pulses

Krotov.print_table

Krotov.KrotovResult — Type

Result object returned by optimize_pulses.

Krotov.optimize_pulses — Method

Use Krotov's method to optimize the given optimization problem.

result = optimize_pulses(problem)

optimizes the control problem, see QuantumControlBase.ControlProblem.

Parameters are taken from the keyword arguments used in the instantiation of problem.

Required problem keyword arguments

The optimization functional is given implicitly via the mandatory problem keyword argument chi.

Optional problem keyword arguments

The following problem keyword arguments are supported (with default values):

sigma=nothing: Function that calculate the second-order contribution. If not given, the first-order Krotov method is used.

iter_start=0: the initial iteration number

iter_stop=5000: the maximum iteration number

check_convergence: a function to check whether convergence has been reached. Receives a KrotovResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘.

prop_method=:auto: The propagation method to use

Krotov.print_table — Method

Default info_hook