API

Krotov.KrotovResult

Krotov.PulseParametrization

Krotov.LogisticParametrization

Krotov.LogisticSqParametrization

Krotov.NoParametrization

Krotov.SquareParametrization

Krotov.TanhParametrization

Krotov.TanhSqParametrization

Krotov.optimize_pulses

Krotov.print_table

Krotov.KrotovResult — Type

Result object returned by optimize_pulses.

Krotov.PulseParametrization — Type

Specification for a "time-local" pulse parametrization.

The parametrization is given as a collection of three functions:

$ϵ(u(t))$

$u(ϵ(t))$

$∂ϵ/∂u$ as a function of $u(t)$.

Krotov.LogisticParametrization — Method

Parametrization with a Logistic function that enforces ϵmin < ϵ(t) < ϵmax.

Krotov.LogisticSqParametrization — Method

Parametrization with a Logistic-Square function that enforces 0 ≤ ϵ(t) < ϵ_max.

Krotov.NoParametrization — Method

Parametrization where $ϵ(t) ≡ u(t)$.

Krotov.SquareParametrization — Method

Parametrization ϵ(t) = u²(t), enforcing pulse values ≥ 0.

Krotov.TanhParametrization — Method

Parametrization with a tanh function that enforces ϵmin < ϵ(t) < ϵmax.

Krotov.TanhSqParametrization — Method

Parametrization with a tanh² function that enforces $0 ≤ ϵ(t) < ϵ_max$.

Krotov.optimize_pulses — Method

Use Krotov's method to optimize the given optimization problem.

result = optimize_pulses(problem; kwargs...)

optimizes the given control problem, see QuantumControlBase.ControlProblem.

Keyword arguments that control the optimization are taken from the keyword arguments used in the instantiation of problem. Any kwargs passed directly to optimize_pulses will update (overwrite) the parameters in problem.

Required problem keyword arguments

J_T: A function J_T(ϕ, objectives) that evaluates the final time functional from a list ϕ of forward-propagated states and problem.objectives.

chi: A function chi!(χ, ϕ, objectives) what receives a list ϕ of the forward propagates state and must set $χₖ=∂J_T/∂⟨ϕₖ|$.

Optional problem keyword arguments

The following keyword arguments are supported (with default values):

sigma=nothing: Function that calculate the second-order contribution. If not given, the first-order Krotov method is used.

iter_start=0: the initial iteration number

iter_stop=5000: the maximum iteration number

prop_method=:auto: The propagation method to use

update_hook: A function that receives the Krotov workspace, the iteration number, the list of updated pulses and the list of guess pulses as positional arguments. The function may mutate any of its arguments. This may be used e.g. to apply a spectral filter to the updated pulses, or to update propagation workspaces inside the Krotov workspace.

info_hook: A function that receives the same argumens as update_hook, in order to write information about the current iteration to the screen or to a file. The default info_hook prints a table with convergence information to the screen. Runs after update_hook. The info_hook function may return a tuple, which is stored in the list of records inside the KrotovResult object.

check_convergence: a function to check whether convergence has been reached. Receives a KrotovResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘. The convergence check is performed after any calls to update_hook and info_hook.

Krotov.print_table — Method

Default info_hook