API

Krotov.KrotovResult

Krotov.optimize_krotov

Krotov.print_table

QuantumControlBase.optimize

Krotov.KrotovResult — Type

Result object returned by optimize_krotov.

Krotov.optimize_krotov — Method

See optimize(problem; method=:krotov, kwargs...).

Krotov.print_table — Method

Print optimization progress as a table.

This functions serves as the default info_hook for an optimization with Krotov's method.

QuantumControlBase.optimize — Method

result = optimize(problem; method=:krotov, kwargs...)

optimizes the given control problem using Krotov's method, returning a KrotovResult.

Keyword arguments that control the optimization are taken from the keyword arguments used in the instantiation of problem.

Required problem keyword arguments

J_T: A function J_T(ϕ, objectives) that evaluates the final time functional from a list ϕ of forward-propagated states and problem.objectives.

Recommended problem keyword arguments

lambda_a=1.0: The inverse Krotov step width λ_a for every pulse.

update_shape=(t->1.0): A function S(t) for the "update shape" that scales the update for every pulse

If different controls require different lambda_a or update_shape, a dict pulse_options must be given instead of a global lambda_a and update_shape, see below.

Optional problem keyword arguments

The following keyword arguments are supported (with default values):

pulse_options: A dictionary that maps every control (as obtained by get_controls from the problem.objectives) to the following dict:

:lambda_a: The value for inverse Krotov step width λₐ

:update_shape: A function S(t) for the "update shape" that scales the Krotov pulse update.

This overrides the global lambda_a and update_shape arguments.

chi: A function chi!(χ, ϕ, objectives) what receives a list ϕ of the forward propagated states and must set $|χₖ⟩ = -∂J_T/∂⟨ϕₖ|$. If not given, it will be automatically determined from J_T via make_chi with the default parameters.

sigma=nothing: Function that calculate the second-order contribution. If not given, the first-order Krotov method is used.

iter_start=0: the initial iteration number

iter_stop=5000: the maximum iteration number

prop_method/fw_prop_method/bw_prop_method: The propagation method to use for each objective, see below.

update_hook: A function that receives the Krotov workspace, the iteration number, the list of updated pulses and the list of guess pulses as positional arguments. The function may mutate any of its arguments. This may be used e.g. to apply a spectral filter to the updated pulses, or to update propagation workspaces inside the Krotov workspace.

info_hook: A function (or tuple of functions) that receives the same argumens as update_hook, in order to write information about the current iteration to the screen or to a file. The default info_hook prints a table with convergence information to the screen. Runs after update_hook. The info_hook function may return a tuple, which is stored in the list of records inside the KrotovResult object.

check_convergence: a function to check whether convergence has been reached. Receives a KrotovResult object result, and should set result.converged to true and result.message to an appropriate string in case of convergence. Multiple convergence checks can be performed by chaining functions with ∘. The convergence check is performed after any calls to update_hook and info_hook.

verbose=false: If true, print information during initialization

The propagation method for the forward propagation of each objective is determined by the first available item of the following:

a fw_prop_method keyword argument

a prop_method keyword argument

a property fw_prop_method of the objective

a property prop_method of the objective

the value :auto

The propagation method for the backword propagation is determined similarly, but with bw_prop_method instead of fw_prop_method.